change: express endpoints refactored into routes for better organization
