+++ client/app/components/JoinHouseForm/JoinHouseForm.jsx
@@ -11,7 +11,7 @@ export default function JoinHouseForm({userData}) {
             houseId:  document.getElementById("houseId").value,
         }
 
-        fetch(`http://localhost:2001/joinHouse?userId=${userData.id}`, {
+        fetch(`http://localhost:2001/houses/join?userId=${userData.id}`, {
             method: 'POST',
             headers: {'Content-Type' : 'application/json'},
             body: JSON.stringify(formBody)
