+++ client/app/js/AuthContext.js
@@ -24,7 +24,7 @@ export function AuthContextProvider({ children }) {
     useEffect(() => {
         if (user) {
             // create user in database if they are not already there
-            fetch('http://localhost:2001/createUser', {method: 'POST', headers: {"Content-Type": "application/json"},
+            fetch('http://localhost:2001/users', {method: 'POST', headers: {"Content-Type": "application/json"},
             body: JSON.stringify(user)})
             .then(() => user.getIdToken().then((token) => {
                 fetch(`http://localhost:2001/users/${user.uid}`, {headers: {Authorization : `Bearer ${token}`}})
